requirements.txt
pip install ____ :

pygame
neat-python

